requirements.txt
Package                  Version     
------------------------ ------------
absl-py                  0.9.0       
astor                    0.8.1       
astroid                  2.4.1       
audioread                2.1.8       
autopep8                 1.5.3       
backcall                 0.2.0       
cachetools               4.1.1       
certifi                  2020.6.20   
cffi                     1.14.0      
chardet                  3.0.4       
cloudpickle              1.3.0       
configobj                5.0.6       
contextlib2              0.6.0.post1 
cycler                   0.10.0      
Cython                   0.29.20     
DateTime                 4.3         
decorator                4.4.2       
ffmpeg-python            0.2.0       
future                   0.18.2      
gast                     0.3.3       
google-api-core          1.21.0      
google-api-python-client 1.10.0      
google-auth              1.19.2      
google-auth-httplib2     0.0.4       
google-pasta             0.2.0       
googleapis-common-protos 1.52.0      
grpcio                   1.27.2      
gym                      0.17.2      
h5py                     2.10.0      
httplib2                 0.18.1      
idna                     2.10        
imageio                  2.8.0       
imutils                  0.5.3       
int-date                 0.1.8       
ipython                  7.16.1      
ipython-genutils         0.2.0       
isort                    4.3.21      
jedi                     0.17.1      
joblib                   0.14.1      
Keras                    2.2.4       
Keras-Applications       1.0.8       
Keras-Preprocessing      1.1.2       
keras-vggface            0.6         
kiwisolver               1.1.0       
lazy-object-proxy        1.4.3       
librosa                  0.7.2       
llvmlite                 0.31.0      
lxml                     4.5.1       
Markdown                 3.2.1       
matplotlib               3.2.1       
mccabe                   0.6.1       
mock                     4.0.2       
networkx                 2.4         
numba                    0.48.0      
numpy                    1.16.0      
object-detection         0.1         
opencv-python            4.2.0.32    
pandas                   0.24.2      
parso                    0.7.0       
pexpect                  4.8.0       
pickleshare              0.7.5       
Pillow                   7.1.1       
pip                      10.0.1      
pomegranate              0.13.3      
prompt-toolkit           3.0.5       
protobuf                 3.12.2      
ptyprocess               0.6.0       
pyasn1                   0.4.8       
pyasn1-modules           0.2.8       
PyAudio                  0.2.11      
pycocotools              2.0.1       
pycodestyle              2.6.0       
pycparser                2.20        
pyglet                   1.5.0       
Pygments                 2.6.1       
pygraphviz               1.5         
pylint                   2.5.2       
pynput                   1.6.8       
pyparsing                2.4.6       
python-dateutil          2.8.1       
python-vlc               3.0.10114   
python-xlib              0.27        
pytz                     2020.1      
PyWavelets               1.1.1       
PyYAML                   5.3.1       
requests                 2.24.0      
resampy                  0.2.2       
rsa                      4.6         
scikit-image             0.16.2      
scikit-learn             0.22.2.post1
scipy                    1.4.1       
seaborn                  0.10.1      
setuptools               46.1.3      
six                      1.14.0      
slim                     0.1         
SoundFile                0.10.3.post1
SpeechRecognition        3.8.1       
stockstats               0.3.1       
tensorboard              1.14.0      
tensorflow-estimator     1.14.0      
tensorflow-gpu           1.14.0      
termcolor                1.1.0       
tf-slim                  1.1.0       
toml                     0.10.1      
tqdm                     4.45.0      
traitlets                4.3.3       
typed-ast                1.4.1       
uritemplate              3.0.1       
urllib3                  1.25.9      
wcwidth                  0.2.5       
Werkzeug                 1.0.0       
wheel                    0.34.2      
wrapt                    1.12.1      
zope.interface           5.1.0       
